async dnn[1/4]: use activate func for dnn_processing
A patch series including 4 patches are cooked for supporting async
inference for dnn module. This is the kicking off work which implmented
the generic inference filter: dnn_processing with async capability. And
the first backend that supports async inference is implemented in this
patch set.

This patch is the first one of this patch series.